+++ src/hooks/Guilds/contracts/useRichContractRegistry.test.ts
@@ -1,8 +1,8 @@
 import { renderHook } from '@testing-library/react-hooks';
 import {
   IPFSRichContractData,
-  useRichContractData,
-} from './useRichContractData';
+  useRichContractRegistry,
+} from './useRichContractRegistry';
 
 const mockChainId = 1337;
 const contractAddress = '0x0000000000000000000000000000000000000001';
@@ -28,9 +28,9 @@ jest.mock('hooks/Guilds/ipfs/useIPFSFile', () => ({
   }),
 }));
 
-describe('useRichContractData', () => {
+describe('useRichContractRegistry', () => {
   it('should return correct contract data', async () => {
-    const { result } = renderHook(() => useRichContractData());
+    const { result } = renderHook(() => useRichContractRegistry());
 
     expect(result.current.contracts).not.toBe(null);
     expect(result.current.contracts.length).toBe(1);
