+++ src/old-components/Guilds/ActionsBuilder/types.ts
@@ -1,6 +1,6 @@
 import { BigNumber } from 'ethers';
 import { utils } from 'ethers';
-import { RichContractData } from 'hooks/Guilds/contracts/useRichContractData';
+import { RichContractData } from 'hooks/Guilds/contracts/useRichContractRegistry';
 
 export enum SupportedAction {
   ERC20_TRANSFER = 'ERC20_TRANSFER',
